2024.6
Updates

* Updated `fmt` to resolve some warnings occurring with newer compiler versions. Migrated to using custom fork to resolve an issue on platforms that are not fully posix-compliant
* Refactor of internal serialization functionality for `PagedResult`
* Explicit Alphabetical sort ordering for `FilterParams`

Bugfixes

* Fix for a race condition in accessing `SDKSessionData`

Breaking Changes

* `Modio::ModInfo::Price` field is now `uint64_t` for consistency

+++ modio/modio/core/entities/ModioModInfo.h
@@ -133,7 +133,7 @@ namespace Modio
 		Modio::ObjectVisibility Visibility = Modio::ObjectVisibility::Public;
 
 		/// @brief Price of this mod
-		int Price = 0;
+		uint64_t Price = 0;
 
 		/// @brief If this mod has any direct dependencie.
 		bool Dependencies;
